move ServiceProvider and storage stuff to a separate crate
