+++ src/lib.rs
@@ -1,5 +1,6 @@
-pub mod constants;
 pub mod init_tunnel;
 mod utils;
 pub mod types;
 pub mod fetch_api;
+mod storage;
+pub mod constants;
