+++ src/types/request.rs
@@ -4,8 +4,9 @@ use std::collections::HashMap;
 use ntor::common::NTorParty;
 use web_sys::{ReferrerPolicy, RequestMode};
 use web_sys::{AbortSignal, console, Request, RequestInit, ResponseInit};
+use crate::storage::DEV_FLAG;
 use crate::types::response::L8ResponseObject;
-use crate::types::{Body, Mode, network_state::{DEV_FLAG, NetworkStateOpen}, NetworkResponse, WasmEncryptedMessage};
+use crate::types::{Body, Mode, network_state::{NetworkStateOpen}, NetworkResponse, WasmEncryptedMessage};
 use crate::utils;
 
 /// A JSON serializable wrapper for a request that can be sent using the Fetch API.
